refactor(frontend): move formatBytes to shared utils
