+++ frontend/lib/utils.ts
@@ -8,3 +8,11 @@ export function cn(...inputs: ClassValue[]) {
 export function getChatStorageKey(projectId: string): string {
   return `chat-history-${projectId}`;
 }
+
+export function formatBytes(bytes: number) {
+  if (bytes === 0) return "0 B";
+  const k = 1024;
+  const sizes = ["B", "KB", "MB", "GB", "TB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${Number.parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`;
+}
